Fix PSR1 violations
